refactor(playground): modularize FileList component
